Add Feed for users.
Feed is available for each user on the user's page.
The feed for all subscriptions is available on /feed link.
On this pages only 10 latest messages are shown.
Each access to /feed page is cached for 1 minute and served from redis.

Each new message is saved in the database and sent to rabbitmq.
Consumer command reads id of the message from rabbitmq and saves the link
from message to subscriber in database shards.